pyproject.toml
[project]
name = "api-contract-env"
version = "1.0.0"
description = "API Contract Negotiator — OpenEnv RL Environment"
requires-python = ">=3.10"

dependencies = [
    "openenv-core",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
